refactor: split vault process observation helpers
